+++ eggs/Naaya/eggs/Naaya/CHANGELOG.rst
@@ -1,8 +1,11 @@
-2.12.31 (unreleased)
+2.12.32 (unreleased)
+--------------------
+* gl_changeLanguage properly redirects when no referer [simiamih]
+
+2.12.31 (2011-11-17)
 --------------------
 * portal_map methods are no longer called if the content type is not
   geo_enabled [dumitval]
-* gl_changeLanguage properly redirects when no referer [simiamih]
 
 2.12.30 (2011-11-16)
 --------------------